import fixes, modify moved to ast
