+++ packages/graphql-lowcode/src/index.ts
@@ -1,5 +1,5 @@
 import * as generate from './generate'
-import * as modify from './modify'
+import * as modify from './ast/modify'
 import * as ast from './ast'
 
 export { generate, modify, ast }
